small fix and improve doc

+++ enot/action/release.py
@@ -1,11 +1,13 @@
+import fileinput
 import shutil
 import subprocess
-from os import remove
+import sys
+from os import remove, listdir
 from os.path import join
 
 from enot.action.action import Action
 from enot.compiler.relx import RelxCompiler
-from enot.utils.file_utils import untar, remove_dir
+from enot.utils.file_utils import untar, remove_dir, ensure_executable
 from enot.utils.logger import error
 
 
@@ -23,10 +25,11 @@ class Release(Action):
         rel_dir = self.params['rel_dir']
         erts_tarball = None
         try:
-            erts_tarball = self.__download_erts(enot_cache, package.path, erlang_vsn)
+            erts_tarball, erts_path = Release.__download_erts(enot_cache, package.path, erlang_vsn)
             compiler = RelxCompiler(package)
             compiler.ensure_tool(system_config.cache.local_cache)
-            compiler.compile(params=['-i', package.path])
+            if not compiler.compile(params=['-i', package.path], erts=erts_path):
+                return False
             local = join(rel_dir, '_rel')
             remove_dir(local)
             shutil.copytree(join(package.path, '_rel'), local)
@@ -44,9 +47,30 @@ class Release(Action):
     def export(self) -> dict:
         return {'release': self.params}
 
-    def __download_erts(self, enot_cache, path, erlang_vsn):
+    @staticmethod
+    def __download_erts(enot_cache, path, erlang_vsn):
         if enot_cache is None:
             raise RuntimeError('No official enot cache in config')
         downloaded_path = enot_cache.fetch_erts(erlang_vsn)
         untar(downloaded_path, path)
-        return downloaded_path
+        erts_path = Release.__determine_erts(path)
+        Release.__change_erts_root(erts_path, path)
+        return downloaded_path, erts_path
+
+    @staticmethod
+    def __determine_erts(path):
+        erts = [f for f in listdir(path) if f.startswith("erts-")]
+        if not erts:
+            raise RuntimeError('Erts not found in ' + path)
+        erts_bin = join(path, erts[0], 'bin')
+        for f in listdir(erts_bin):  # make all files executable so relx can run
+            ensure_executable(join(erts_bin, f))
+        return erts_bin
+
+    @staticmethod
+    def __change_erts_root(erts, path):
+        for i, line in enumerate(fileinput.input(join(erts, 'erl'), inplace=True)):
+            if line.startswith('ROOTDIR="//erl/'):
+                sys.stdout.write('ROOTDIR="' + path + '"\n')
+            else:
+                sys.stdout.write(line)
